work on breaking out simplifiers into own module
