+++ wowdump/simplifiers/test.py
@@ -1,0 +1,6 @@
+def fcall():
+    print("meow")
+
+map = [
+    ("regex", fcall)
+]
